third_party/WebKit/LayoutTests/svg/dynamic-updates/SVGEllipseElement-svgdom-requiredFeatures-expected.txt

SVG 1.1 dynamic update tests

Check that SVGEllipseElement is initially displayed
PASS ellipseElement.getBBox().width is 400
PASS ellipseElement.getBBox().height is 400
Check that setting requiredFeatures to something invalid makes it not render
PASS ellipseElement.getBBox().width is 0
PASS ellipseElement.getBBox().height is 0
Check that setting requiredFeatures to something valid makes it render again
PASS ellipseElement.getBBox().width is 400
PASS ellipseElement.getBBox().height is 400
Check that adding something valid to requiredFeatures keeps rendering the element
PASS ellipseElement.getBBox().width is 400
PASS ellipseElement.getBBox().height is 400
Check that adding something invalid to requiredFeatures makes it not render
PASS ellipseElement.getBBox().width is 0
PASS ellipseElement.getBBox().height is 0
PASS successfullyParsed is true

TEST COMPLETE
